alvirity_site/alvirityru/requirements.txt
asgiref==3.3.1
certifi==2021.5.30
chardet==4.0.0
Django==3.1.5
django-ckeditor==5.6.1
django-js-asset==1.2.2
django-modeltranslation==0.15.2
gunicorn==20.0.4
idna==2.10
pytils==0.4.1
pytz==2020.5
requests==2.25.1
setuptools==44.1.1
six==1.16.0
sqlparse==0.4.1
Unidecode==1.2.0
urllib3==1.26.5
wheel==0.34.2